Update simdjson to support 64 bit integers
Update to the latest simdjson version, and check for 64 bits
signed/unsigned integers. Unfortunately, there is no unsigned version of
int8_numeric, so we need to do a bit of transformations there, namely
flip the top bit, convert the value as if signed, subtract -(2^63) from
the result (thanks to Andrew Gierth for suggestion).